实践36096-对ST分区均衡竞争算法中的防抽具低值的进行降权。

+++ SMG_NothingIsAll/AIFoundation/AIThinkingControl/TCRecognition/Model/AIFeatureJvBuModel.h
@@ -47,7 +47,7 @@
 -(NSString*) getSTMatchDesc;
 -(NSString*) getGTMatchDesc;
 
-@property (assign, nonatomic) NSInteger areaRankSum; // 考试名次之和（越大越好）。
+@property (assign, nonatomic) CGFloat areaRankSum; // 考试名次之和（越大越好）。
 @property (assign, nonatomic) NSInteger areaRankNum; // 考试次数（用于计算均排名）。
 -(CGFloat) areaRankScore; // 平均名次（越大越好）。
 -(void) run4ItemAreaRankScore:(NSArray*)stModels;
